Create dir for Parameters and extract util functions
